Migrated QueryHooks for Categories/Tags
