--- layers/Schema/packages/migrate-post-tags-wp/migrate/plugins/load.php
@@ -1,5 +1,0 @@
-<?php
-
-if (class_exists('\PoPSchema\PostsWP\Component')) {
-	require_once 'pop-posts-wp/load.php';
-}
