--- layers/Schema/packages/migrate-post-tags-wp/migrate/plugins/pop-posts-wp/kernel/load.php
@@ -1,4 +1,0 @@
-<?php
-
-require_once 'cms-functionapi-hooks.php';
-require_once 'cms-functionapi-hooks-categories.php';
