--- layers/Schema/packages/migrate-post-tags-wp/migrate/plugins/pop-posts-wp/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'kernel/load.php';
